requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
filelock==3.20.1
fsspec==2025.12.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sentencepiece==0.2.1
sympy==1.14.0
tokenizers==0.15.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.39.3
typing_extensions==4.15.0
urllib3==2.6.2
